+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.3.0-rc.35",
-  genVersion: "2.809.2",
-  userAgent: "speakeasy-sdk/typescript 4.3.0-rc.35 2.809.2 2.0 @orq-ai/node",
+  sdkVersion: "4.3.0-rc.37",
+  genVersion: "2.811.4",
+  userAgent: "speakeasy-sdk/typescript 4.3.0-rc.37 2.811.4 2.0 @orq-ai/node",
 } as const;