+++ packages/orq-rc/src/models/components/conversationwithmessagesresponse.ts
@@ -122,7 +122,7 @@ export const ConversationWithMessagesResponse$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kgsj2jk5xy8xttrbt7d2kj0w"),
+  _id: z.string().default("conv_01kgysbcgy980e5qa7gr57mvn9"),
   kind: ConversationWithMessagesResponseKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),