+++ packages/orq-rc/src/models/operations/createconversationresponse.ts
@@ -280,7 +280,7 @@ export const PartReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01kgsj2jkvjhdyvv5sn10ewkms"),
+  _id: z.string().default("reasoning_01kgysbchj9tgk55wy98b21wp6"),
   metadata: z.record(z.any()).optional(),
   kind: PartKind$inboundSchema,
   reasoning: z.string(),
@@ -366,7 +366,7 @@ export const ReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01kgsj2jkpjn7hy1vacdybwfzd"),
+  _id: z.string().default("reasoning_01kgysbchevphxb4jmgm057peg"),
   metadata: z.record(z.any()).optional(),
   kind: DeltaKind$inboundSchema,
   reasoning: z.string(),