+++ packages/orq-rc/src/models/operations/updateconversation.ts
@@ -226,7 +226,7 @@ export const UpdateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kgsj2jkc5adxcaj078k7ad23"),
+  _id: z.string().default("conv_01kgysbch4pqbknf7hj1w66xwf"),
   kind: UpdateConversationKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),